Abstracted transport layer to an independent module
This also adds back the option to use the i2c interface with the i2c
feature.
