+++ src/ecc.rs
@@ -1,33 +1,30 @@
 use crate::command::EccError;
-use crate::constants::{ATCA_CMD_SIZE_MAX, WAKE_DELAY};
+use crate::constants::ATCA_CMD_SIZE_MAX;
+use crate::transport::EccTransport;
 use crate::{
     command::{EccCommand, EccResponse},
     Address, DataBuffer, Error, KeyConfig, Result, SlotConfig, Zone,
 };
 use bytes::{BufMut, Bytes, BytesMut};
-use serialport::{ClearBuffer, DataBits, SerialPort, StopBits};
 use sha2::{Digest, Sha256};
 use std::{thread, time::Duration};
 
 pub use crate::command::KeyType;
 
 pub struct Ecc {
-    port: String,
+    transport: EccTransport,
 }
 
 pub const MAX_SLOT: u8 = 15;
 
-pub(crate) const RECV_RETRIES: u8 = 2;
-pub(crate) const RECV_RETRY_WAIT: Duration = Duration::from_millis(50);
-pub(crate) const CMD_RETRIES: u8 = 10; 
+pub(crate) const CMD_RETRIES: u8 = 10;
 
 impl Ecc {
     pub fn from_path(path: &str, address: u16) -> Result<Self> {
 
-        let _ = address; //keep the API the same. Address refers to i2c addr which isn't required for SWI
-        let port = String::from( path );
+        let transport = EccTransport::from_path( path, address)?;
 
-        Ok(Self {port})
+        Ok(Self {transport})
     }
 
     pub fn get_info(&mut self) -> Result<Bytes> {
@@ -127,7 +124,7 @@ impl Ecc {
             match response{
                 Ok(_) => (),
                 Err(_) if attempts < 3 => {
-                    self.send_sleep();
+                    self.transport.send_sleep();
                     continue;
                 }
                 Err(_) => return response,
@@ -141,7 +138,7 @@ impl Ecc {
             match response{
                 Ok(_) => return response,
                 Err(_) if attempts < 3 => {
-                    self.send_sleep();
+                    self.transport.send_sleep();
                     continue;
                 }
                 Err(_) => return response,
@@ -177,83 +174,6 @@ impl Ecc {
             .map(|_| ())
     }
 
-    fn send_wake(&mut self) -> Result {
-        let port_name = &self.port;
-        let baud_rate = 115_200;
-        let stop_bits = StopBits::One;
-        let data_bits = DataBits::Eight;
-        let uart_wake_builder = serialport::new(port_name, baud_rate)
-            .stop_bits(stop_bits)
-            .data_bits(data_bits);
-
-        let mut uart_wake = uart_wake_builder.open().unwrap_or_else(|e| {
-            eprintln!("Failed to open port {}. Error: {}", port_name,e);
-            ::std::process::exit(1);
-        });
-        let _ = uart_wake.write(&[0]);
-        
-        thread::sleep(WAKE_DELAY);
-        self.read_wake_response()
-    }
-
-    fn read_wake_response( &mut self) -> Result {
-        let port_name = &self.port;
-        let baud_rate = 230_400;
-        let stop_bits = StopBits::One;
-        let data_bits = DataBits::Seven;
-        let uart_cmd_builder = serialport::new(port_name, baud_rate)
-            .stop_bits(stop_bits)
-            .data_bits(data_bits);
-
-        let mut uart_cmd = uart_cmd_builder.open().unwrap_or_else(|e| {
-            eprintln!("Failed to open port {}. Error: {}", port_name,e);
-            ::std::process::exit(1);
-        });
-        
-        // Send transmit flag to signal bus
-        let mut transmit_flag = BytesMut::new();
-        transmit_flag.put_u8(0x88);
-        let encoded_transmit_flag = self.encode_uart_to_swi(&transmit_flag );
-        uart_cmd.write(&encoded_transmit_flag)?;
-        thread::sleep(Duration::from_micros(5_000) );
-        
-        let mut encoded_msg = BytesMut::new();
-        encoded_msg.resize(40,0);
-        let _ = uart_cmd.read(&mut encoded_msg);
-
-        let mut decoded_msg = BytesMut::new();
-        decoded_msg.resize(5, 0);
-        
-        self.decode_swi_to_uart(&encoded_msg, &mut decoded_msg);
-        
-        let response = EccResponse::from_bytes(&decoded_msg[1..]);
-        match response {
-            Err(e) => return Err(e),
-            _ => return Ok(()),
-        }
-    }
-
-    fn send_sleep(&mut self) {        
-        let port_name = &self.port;
-        let baud_rate = 230_400;
-        let stop_bits = StopBits::One;
-        let data_bits = DataBits::Seven;
-        let uart_cmd_builder = serialport::new(port_name, baud_rate)
-            .stop_bits(stop_bits)
-            .data_bits(data_bits);
-
-        let mut uart_cmd = uart_cmd_builder.open().unwrap_or_else(|e| {
-            eprintln!("Failed to open port {}. Error: {}", port_name,e);
-            ::std::process::exit(1);
-        });
-
-        let mut sleep_msg = BytesMut::new();
-        sleep_msg.put_u8(0xCC);
-        let sleep_encoded = self.encode_uart_to_swi(&sleep_msg);
-
-        let _ = uart_cmd.write(&sleep_encoded);
-    }
-
     pub(crate) fn send_command(&mut self, command: &EccCommand) -> Result<Bytes> {
         self.send_command_retries(command, true, CMD_RETRIES)
     }
@@ -266,7 +186,7 @@ impl Ecc {
     ) -> Result<Bytes> {
         let mut buf = BytesMut::with_capacity(ATCA_CMD_SIZE_MAX as usize);
         for retry in 0..retries {
-            let response = self.send_wake();
+            let response = self.transport.send_wake();
             
             match response {
                 Ok(_) => (),
@@ -282,7 +202,7 @@ impl Ecc {
             buf.clear();         
             command.bytes_into(&mut buf);
             
-            if let Err(_) = self.send_recv_buf(command.duration(), &mut buf){
+            if let Err(_) = self.transport.send_recv_buf(command.duration(), &mut buf){
                 if retry == retries {
                     break;
                 } else {
@@ -293,7 +213,7 @@ impl Ecc {
             
             let response = EccResponse::from_bytes(&buf[..]);
             if sleep {
-                self.send_sleep();
+                self.transport.send_sleep();
             }
             match response {
                 Ok(EccResponse::Data(bytes)) => return Ok(bytes),
@@ -314,129 +234,5 @@ impl Ecc {
             }
         }
         Err(Error::timeout())
-    }
-
-    fn send_recv_buf(&mut self, delay: Duration, buf: &mut BytesMut) -> Result {
-        
-        let port_name = &self.port;
-        let baud_rate = 230_400;
-        let stop_bits = StopBits::One;
-        let data_bits = DataBits::Seven;
-        let uart_cmd_builder = serialport::new(port_name, baud_rate)
-            .stop_bits(stop_bits)
-            .data_bits(data_bits);
-
-        let mut uart_driver = uart_cmd_builder.open().unwrap_or_else(|e| {
-            eprintln!("Failed to open port {}. Error: {}", port_name,e);
-            ::std::process::exit(1);
-        });
-        
-        let _ = uart_driver.clear(ClearBuffer::All);
-        let swi_msg = self.encode_uart_to_swi(buf);
-        self.send_buf(&swi_msg, &mut uart_driver)?;
-        thread::sleep(delay);
-        self.recv_buf(buf, &mut uart_driver)
-    }
-
-    pub(crate) fn send_buf(&mut self, buf: &[u8], serial_port: &mut Box<dyn SerialPort>) -> Result {
-        
-        let send_size = serial_port.write(buf)?;
-
-        //Each byte takes ~45us to transmit, so we must wait for the transmission to finish before proceeding
-        let uart_tx_time = Duration::from_micros( (buf.len() * 45) as u64); 
-        thread::sleep(uart_tx_time);
-        //Because Tx line is linked with Rx line, all sent msgs are returned on the Rx line and must be cleared from the buffer
-        let mut clear_rx_line = BytesMut::new();
-        clear_rx_line.resize(send_size, 0);
-        let _ = serial_port.read_exact( &mut clear_rx_line );
-
-        Ok(())
-    }
-
-    pub(crate) fn recv_buf(&mut self, buf: &mut BytesMut,  serial_port: &mut Box<dyn SerialPort>) -> Result {
-        let mut encoded_msg = BytesMut::new();
-        encoded_msg.resize(ATCA_CMD_SIZE_MAX as usize,0);
-        
-        let mut transmit_flag = BytesMut::new();
-        transmit_flag.put_u8(0x88);
-        let encoded_transmit_flag = self.encode_uart_to_swi(&transmit_flag );
-        
-        let _ = serial_port.clear(ClearBuffer::All);
-
-        for retry in 0..RECV_RETRIES {
-            serial_port.write(&encoded_transmit_flag)?;
-            thread::sleep(Duration::from_micros(40_000) );
-            let read_response = serial_port.read(&mut encoded_msg);
-            
-            match read_response {
-                Ok(cnt) if cnt == 8 => { //If the buffer is empty except for the transmit flag, wait & try again
-                },
-                Ok(cnt) if cnt > 16 => {
-                    break;
-                },
-                _ if retry != RECV_RETRIES => continue,
-                _  => return Err(Error::Timeout) 
-            }
-            
-            thread::sleep(RECV_RETRY_WAIT);
-        }
-
-        let mut decoded_message = BytesMut::new();
-        decoded_message.resize((ATCA_CMD_SIZE_MAX) as usize, 0);   
-
-        self.decode_swi_to_uart(&encoded_msg, &mut decoded_message);
-
-        let encoded_msg_size = decoded_message[1];
-
-        if encoded_msg_size as u16 > ATCA_CMD_SIZE_MAX/8{
-            return Err(Error::Timeout)
-        }
-
-        buf.resize(encoded_msg_size as usize, 0);
-
-        // Remove the transmit flag at the beginning & the excess buffer space at the end
-        let _transmit_flag = decoded_message.split_to(1);
-        decoded_message.truncate(encoded_msg_size as usize);
-
-        buf.copy_from_slice(&decoded_message);
-
-        Ok(())
-    }
-
-    fn encode_uart_to_swi(&mut self, uart_msg: &BytesMut ) -> BytesMut {
-        
-        let mut bit_field = BytesMut::new();
-        bit_field.reserve(uart_msg.len() * 8 );
-    
-        for byte in uart_msg.iter() {
-            for bit_index in 0..8 {
-                if ( ((1 << bit_index ) & byte) >> bit_index ) == 0 {
-                    bit_field.put_u8(0xFD); 
-                } else {
-                    bit_field.put_u8(0xFF);
-                }
-            }
-        }
-        bit_field
-    }
-    
-    fn decode_swi_to_uart(&mut self, swi_msg: &BytesMut, uart_msg: &mut BytesMut ) {
-    
-        uart_msg.clear();
-        assert!( (swi_msg.len() % 8) == 0);
-        uart_msg.resize( &swi_msg.len() / 8, 0 );
-    
-        let mut i = 0; 
-        for byte in uart_msg.iter_mut() {
-            let bit_slice= &swi_msg[i..i+8];
-            
-            for bit in bit_slice.iter(){
-                if *bit == 0x7F || *bit == 0x7E {
-                    *byte ^= 1;
-                }
-                *byte = byte.rotate_right(1);
-            }
-            i += 8;
-        }
     }
 }
